# Garage Management System/src/app/authentication/docs/swagger.yaml
definitions:
  models.BFFCreateCustomerRequest:
    properties:
      confirmPassword:
        example: Admin@123
        minLength: 8
        type: string
      customerName:
        example: Arjit
        maxLength: 32
        minLength: 5
        type: string
      email:
        example: arijit@gmail.com
        type: string
      password:
        example: Admin@123
        maxLength: 20
        minLength: 8
        type: string
      phoneNumber:
        example: 8432805566
        maximum: 9999999999
        minimum: 1000000000
        type: integer
    required:
    - confirmPassword
    - customerName
    - email
    - password
    - phoneNumber
    type: object
  models.ErrorAPIResponse:
    properties:
      error:
        type: string
      errors:
        $ref: '#/definitions/models.ErrorMessage'
    type: object
  models.ErrorMessage:
    properties:
      errorMessage:
        type: string
      key:
        type: string
    type: object
info:
  contact: {}
  description: Authentication APIs for Garage Management System
  title: Authentication Service API
  version: "1.0"
paths:
  /api/auth/register/customer:
    post:
      consumes:
      - application/json
      description: Handles customer registration by validating input and storing user
        details
      parameters:
      - description: Customer Registration Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BFFCreateCustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Customer created successfully
          schema:
            type: string
        "400":
          description: Invalid input payload
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
      summary: Create a new customer
      tags:
      - Customer
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
    x-extension-openapi: '{"example": "value on a json format"}'
swagger: "2.0"